Added test cases for the GL buffer classes and worked on the (still incomplete) iOS GL context implementation.

+++ src/ios/java/fulcrum/ios/opengl/IOSGLBytes.java
@@ -51,7 +51,7 @@ final class IOSGLBytes extends GLBytes {
   }
 
   /** Acquires the appropriate range in the underlying buffer. */
-  int pointer() {
+  long pointer() {
     return buffer.bytePointer(beginIndex);
   }
